@@ -3,13 +3,13 @@
     "description": "GMO Payment Gateway API for simple credit card payments",
     "type": "library",
     "require": {
-        "php": ">=5.6.30",
+        "php": "^7.0",
         "guzzlehttp/guzzle": "^6",
         "sanmai/pipeline": "^0.3.2|^2.0|^3.0.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "<6",
-        "php-coveralls/php-coveralls": "^1.0",
+        "phpunit/phpunit": "^6",
+        "php-coveralls/php-coveralls": ">=1.0",
         "friendsofphp/php-cs-fixer": "^2.6"
     },
     "license": "MIT",
