@@ -23,7 +23,7 @@
 
 namespace GMO\Tests;
 
-class TestCase extends \PHPUnit_Framework_TestCase
+class TestCase extends \PHPUnit\Framework\TestCase
 {
     public static function setUpBeforeClass()
     {
